requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==4.0.5
django-ckeditor==6.4.2
django-filter==22.1
django-js-asset==2.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djangorestframework-xml==2.0.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.1.1
PyJWT==2.4.0
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
